+++ frontend/src/components/chat/dm/ChatFriendsComponent.tsx
@@ -7,6 +7,7 @@ import '../../../styles/Chat.css'
 import { ChatRoom } from "../../../types/PrismaType";
 import LeaveButton from "../utils/LeaveButton";
 import getUserName from "../../../utils/getUserName";
+import { getFriendNameFromRoomName } from "../../../utils/ChatAxios";
 
 
 type ChatFriendsComponentProps = {
@@ -24,18 +25,6 @@ export default function ChatFriendsComponent(props: ChatFriendsComponentProps) {
   const roomID = useLocation().pathname.split('/')[3];
   const [prevRoomId, setPrevRoomId] = useState<string>();
   const [Loading, setLoading] = useState(true);
-
-  const getFriendNameFromRoomName = (room: string, username: string): string => {
-    let friendName: string = '';
-
-    const names = room.split(',');
-    names.map((name) => {
-      if (name !== username) {
-        friendName = name;
-      }
-    })
-    return friendName;
-  }
 
   const getUserDM = async () => {
     const { data } = await axios.get(`http://localhost:8080/chat/dm`);
