+++ frontend/src/components/chat/utils/UserParticipantsList.tsx
@@ -4,6 +4,7 @@ import { Link } from "react-router-dom";
 import React, {useEffect, useState } from "react";
 import InvitationButton from "../group/InvitationButton";
 import AdminButton from "./AdminButton";
+import { getUserId } from "../../../utils/ChatAxios";
 
 type MemberPayload = {
   id: string;
@@ -27,14 +28,7 @@ export default function UserParticipantList(props: UserParticipantListProps) {
   const [Loading, setLoading] = useState(true);
 
   useEffect(() => {
-    const getUserId = async () => {
-      const { data } = await axios.get(`http://localhost:8080/user`);
-      if (data){
-        setUserId(data.id);
-      }
-    }
-
-    getUserId();
+    getUserId().then((id: any) => {setUserId(id);});
   }, [roomId]);
 
   useEffect(() => {
